add changes

+++ src/lambda_function.py
@@ -3,6 +3,7 @@ import json
 
 sts = boto3.client('sts')
 s3 = boto3.resource('s3')
+ec2 = boto3.client('ec2')
 
 def is_self_invocation(detail):
     try:
@@ -16,8 +17,16 @@ def is_self_invocation(detail):
 
     return False
 
-def lambda_handler(event, context):
+def get_bucket_name(instance_id):
+    response = ec2.describe_instances(InstanceIds=[instance_id])
+    instances = response['Reservations'][0]['Instances']
+    for instance in instances:
+        for tag in instance['Tags']:
+            if tag['Key'] == 'Bucket':
+                return tag['Value']
+    return None
 
+def lambda_handler(event, context):
     if 'detail' in event:
         detail = event['detail']
         if is_self_invocation(detail):
@@ -30,23 +39,22 @@ def lambda_handler(event, context):
                 if detail['requestParameters']['eventName'] == 'TerminateInstances':
                     instance_ids = [i['instanceId'] for i in detail['resources']]
                     for instance_id in instance_ids:
-                        bucket_name = 'aws-lambda-bucket-janiator'  # Replace with the name of your S3 bucket
-                        bucket = s3.Bucket(bucket_name)
-                        objects = list(bucket.objects.all())
-                        
-                        if objects:
-                            # Clearing the S3 bucket if it contains objects
-                            bucket.delete_objects(
-                                Delete={
-                                    'Objects': [{'Key': obj.key} for obj in objects]
-                                }
-                            )
-                            print(f"Removed objects from the S3 bucket {bucket_name}")
-                            
-                        # Deleting an S3 bucket if it is empty
-                        if len(list(bucket.objects.all())) == 0:
-                            bucket.delete()
-                            print(f"S3-bucket {bucket_name} deleted")
+                        bucket_name = get_bucket_name(instance_id)
+                        if bucket_name:
+                            bucket = s3.Bucket(bucket_name)
+                            objects = list(bucket.objects.all())
+                            if objects:
+                                # Clearing the S3 bucket if it contains objects
+                                bucket.delete_objects(
+                                    Delete={
+                                        'Objects': [{'Key': obj.key} for obj in objects]
+                                    }
+                                )
+                                print(f"Removed objects from the S3 bucket {bucket_name}")
+                            # Deleting the S3 bucket if it is empty
+                            if len(list(bucket.objects.all())) == 0:
+                                bucket.delete()
+                                print(f"S3 bucket {bucket_name} deleted")
 
     return json.dumps({
         'result': 'SUCCESS',
